requirements.txt
psycopg2-binary==2.8.5
pyyaml==5.3.1
Flask==2.0.3
PyJWT==1.6.4
Flask-JWT-Extended==3.24.1
jsonschema==3.2.0
pycryptodome==3.9.7
Werkzeug~=2.0.0 
